% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance}
\alias{abundance}
\alias{prevalence}
\title{Posterior Samples of Abundance and Prevalence Model Components}
\usage{
abundance(fit, newData, subset = 1:nrow(newData), na.action, breakdown = FALSE)

prevalence(fit, newData, subset = 1:nrow(newData), na.action,
  breakdown = FALSE)
}
\arguments{
\item{fit}{an \code{"fcs2Fit"} object containing a full \acronym{FCS2} model
fit, as returned from \code{\link{fcs2FitModel}} with \code{runBUGS =
TRUE}.}

\item{newData}{a data frame with surveys as rows and variables as columns. It
should contain all variables required by \code{fit}. If missing, the model
matrix contained within \code{fit} is used.}

\item{subset}{an optional vector specifying a subset of surveys to calculate
samples for.}

\item{na.action}{a function which indicates what should happen when the data
contain missing values (\code{NA}s). The default is set by the
\code{na.action} setting of \code{\link{options}} and this is usually set
to \code{\link{na.omit}}. This setting removes surveys that contain missing
data in any required variables. A vector indicating the rows that were
removed can be extracted from the returned object using
\code{\link{na.action}}. Alternatively, \code{\link{na.pass}} can be used
to ignore missing values (where possible) or \code{\link{na.fail}} can be
given to signal an error if missing values are found.}

\item{breakdown}{logical; If \code{FALSE} (the default), samples of the
  abundance or prevalence are returned for each survey. If \code{TRUE}, an
  array is returned contining each additive term in the abundance or
  prevalence regression equation. If \code{breakdown = FALSE} (the default),
  a matrix of posterior samples of the abundance \eqn{\mu} or prevalence
  \eqn{\rho} model component is returned for the selected surveys. Monte
  Carlo samples appear as rows and surveys as columns.

  If \code{breakdown = TRUE}, a three-dimensional array of regression
  component samples is returned. Samples are given in the first dimension,
  surveys in the second and the third dimension corresponds to each additive
  term in the abundance or prevalence regression. The output from
  \code{breakdown = FALSE} can be calculated by summing over the third
  dimension and transforming with \code{\link{exp}} for abundance or
  \code{\link{expit}} for prevalence.

  If surveys were removed by \code{na.action = na.omit}, these can be
  recovered by applying \code{\link{na.action}} to the matrix or array.}
}
\description{
Calculates posterior samples of the abundance or prevalence model components
}
\seealso{
\code{\link{fcs2FitModel}} for producing the required \acronym{FCS2}
  model fit.\cr
}
