% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.array.fcs2EQR.R
\name{as.array.fcs2EQR}
\alias{as.array.fcs2EQR}
\title{Convert EQR object to array}
\usage{
\method{as.array}{fcs2EQR}(x)
}
\arguments{
\item{x}{an object of class \code{"fcs2EQR"} containing Monte Carlo
\acronym{EQR} samples, as calculated from \code{\link{fcs2SingleEQR}},
\code{\link{fcs2JointEQR}} or \code{\link{fcs2JointAndSingleEQR}}.}
}
\value{
Returns the Monte Carlo \acronym{EQR} samples as an
  \code{\link{array}} object.
}
\description{
Converts an \code{"fcs2EQR"} object to a standard \R \code{\link{array}} of
Monte Carlo \acronym{EQR} samples.
}
\seealso{
\code{\link{fcs2SingleEQR}}, \code{\link{fcs2JointEQR}} or
  \code{\link{fcs2JointAndSingleEQR}} for producing \code{"fcs2EQR"}
  objects.\cr \code{[.fcs2EQR} for extracting a subset of
  \acronym{EQR} values.
}
