% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.array.fcs2EQR.R
\encoding{UTF-8}
\name{as.array}
\alias{as.array}
\title{Convert EQR object to array}
\usage{
as.array(x, ...)
}
\arguments{
\item{x}{an object of class \code{"fcs2EQR"} containing Monte Carlo
\acronym{EQR} samples, as calculated from \code{\link{fcs2SingleEQR}},
\code{\link{fcs2JointEQR}} or \code{\link{fcs2JointAndSingleEQR}}.}

\item{\dots}{Not currently used.}
}
\value{
Returns the Monte Carlo \acronym{EQR} samples as an
  \code{\link{array}} object.
}
\description{
Converts an \code{"fcs2EQR"} object to a standard \R \code{\link{array}} of
Monte Carlo \acronym{EQR} samples.
}
\details{
Convert EQR object to array
}
\seealso{
\code{\link{fcs2SingleEQR}}, \code{\link{fcs2JointEQR}} or
  \code{\link{fcs2JointAndSingleEQR}} for producing \code{"fcs2EQR"}
  objects.\cr \code{[.fcs2EQR} for extracting a subset of
  \acronym{EQR} values.
}
