% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClassScot.R
\name{calcClassScot}
\alias{calcClassScot}
\title{Calculate Classification Results for Scotland}
\usage{
calcClassScot(data)
}
\arguments{
\item{data}{dataframe with columns: \describe{ \item{DataHeldBy}{Data owner
descriptor    Field data} \item{SiteCode}{Unique site code e.g. SEPA location
code    Field data} \item{SiteName}{Site name    Field data}
\item{SurveyDate}{Date of survey    Field data} \item{WBId}{SEPA Water body
ID    SEPA GIS system/ SEWeb} \item{WBName}{SEPA Waterbody Name    SEPA GIS
system/ SEWeb} \item{NumberOfRuns}{Number of separate runs in
electrofishing survey    Field data} \item{SurveyArea}{Wetted area of survey
site in square metres    Field data     5-1628} \item{WetWidth}{Wetted width of
survey site in metres    Field data     0.3-41} \item{Slope}{Slope of survey
site, in metres per kilometre    GIS or paper maps    0.01-153}
\item{BarrierType}{Type of impassable barrier downstream, if present. 3
possible values: Artificial, Manmade or Blank    Field data  and/or analysis
of barrier data on SEWeb} \item{ImpassableBarriers}{Is the site above an
impassable barrier? 2 possible values; 1= present, 0= absent    Field data
and/or analysis of barrier data on SEWeb}
\item{CatchmentAreaUpstream}{Total catchment area upstream in square
km    GIS    0.05-2550} \item{CatchmentDrainageDirection}{Catchment drainage
direction. Where does catchment enter the sea? Seven possible categories:
SE= Tweed estuary to Fife Ness E=Fife Ness to Fraserburgh NE= Fraserburgh
to Duncansby Head N= Duncansby Head to Cape Wrath NW= Cape Wrath to Mallaig
W= Mallaig to Mull of Kintyre SW= Mull of Kintyre to Solway Firth}
\item{GeologyClass}{Predominant geology type in catchment. 3 possible
categories: S=Siliceous, C=Calcareous, O=Organic} \item{Altitude}{Site
altitude (metres above sea level)    Field data  or GIS    1-499}
\item{DistanceFromSource}{Distance from the site to the furthest upstream
source (metres)    GIS    115-125066} \item{DistanceToSea}{Distance downstream
from the site to the tidal limit (metres)    GIS    2-160267}
\item{AnnualMeanFlow}{    Annual mean flow in cubic metres per
second    Hydrological modelling (Low Flows Enterprise of similar)
0.003-59.9} \item{AlkalinityValue}{    Alkalinity as CaCO3 (mg/L)     Chemistry
monitoring data    0.54-263.11} \item{TotalPValue}{    Total Phosphorus as P
(mg/L)     Chemistry monitoring data    0.006-0.505} \item{DOCValue}{    Dissolved
organic carbon<0.45μm (mg/L)    Chemistry monitoring data    2.023-24.2}
\item{SuspendedSolidsValue}{    Suspended solids (mg/L)    Chemistry monitoring
data    0.63-32.36} \item{HydrometricAreaNo}{    The SEPA-defined hydrometric
area which the site lies in    SEPA GIS layer}
\item{LandUse.AgriculturalAreas}{% of catchment with Agricultural landuse.
Defined as the following Corrine codes: Pastures (231), Non-irrigated
arable land (211), Complex cultivation patterns (242), Land principally
occupied by agriculture with significant areas or natural vegetation (243)
GIS analysis of Corrine data    0-99.9} \item{LandUse.ConiferousForests}{% of
catchment with coniferous forest landuse (Corrine Code 312)    GIS analysis of
Corrine data    0-92.9} \item{Field}{Description    Source}
\item{LandUse.Wetlands}{% of catchment with wetland landuse. Defined as the
following Corrine codes: Inland marshes (411), Peat bogs (412), Intertidal
flats (423)    GIS analysis of Corrine data    0-95.7}
\item{Substrate.Small}{Total % of survey bed substrate composed of "High
Organic", "silt", "sand" and "gravel", from SFCC survey protocol    Field data
0-100} \item{Substrate.Large}{Total % of survey bed substrate composed of
"Pebble", "cobble" and "boulder" from SFCC survey protocol    Field data
0-100} \item{Substrate.Bedrock}{Total % of survey bed composed of "Bedrock"
from SFCC survey protocol    Field data     0-90}
\item{Salmon_fry.Run1Total}{Total salmon fry (0+) caught in Run 1    Field
data     0-1014} \item{Salmon_fry.Run2Total}{Total salmon fry (0+) caught in
Run 2. Leave blank if no 2nd run fished    Field data     0-352}
\item{Salmon_fry.Run3Total}{Total salmon fry (0+) caught in Run 3. Leave
blank if no 3rd run fished    Field data     0-251}
\item{Salmon_fry.Run4Total}{Total salmon fry (0+) caught in Run 4. Leave
blank if no 4th run fished    Field data     0-12}
\item{Salmon_parr.Run1Total}{Total salmon parr (1++) caught in Run 1    Field
data     0-133} \item{Salmon_parr.Run2Total}{Total salmon parr (1++) caught in
Run 2. Leave blank if no 2nd run fished    Field data     0-68}
\item{Salmon_parr.Run3Total}{Total salmon parr (1++) caught in Run 3. Leave
blank if no 3rd run fished    Field data     0-34}
\item{Salmon_parr.Run4Total}{Total salmon parr (1++) caught in Run 4. Leave
blank if no 4th run fished    Field data     0-15}
\item{Trout_fry.Run1Total}{Total salmon fry (0+) caught in Run 1    Field data
0-508} \item{Trout_fry.Run2Total}{Total salmon fry (0+) caught in Run 2.
Leave blank if no 2nd run fished    Field data     0-225}
\item{Trout_fry.Run3Total}{Total salmon fry (0+) caught in Run 3. Leave
blank if no 3rd run fished    Field data     0-96}
\item{Trout_fry.Run4Total}{Total salmon fry (0+) caught in Run 4. Leave
blank if no 4th run fished    Field data     0-59}
\item{Trout_parr.Run1Total}{Total salmon parr (1++) caught in Run 1    Field
data     0-120} \item{Trout_parr.Run2Total}{Total salmon parr (1++) caught in
Run 2. Leave blank if no 2nd run fished    Field data     0-70}
\item{Trout_parr.Run3Total}{Total salmon parr (1++) caught in Run 3. Leave
blank if no 3rd run fished    Field data     0-43}
\item{Trout_parr.Run4Total}{Total salmon parr (1++) caught in Run 4. Leave
blank if no 4th run fished    Field data     0-25}
}}
}
\value{
dataframe
}
\description{
R Script for calculating new EQRs (for latest FCS2 model) for Scotland
}
\examples{
  \dontrun{
  results <- calcClassScot(data = fcs2::demo_data)#'
}
}
