% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineBUGSFits.R
\name{combineBUGSFits}
\alias{combineBUGSFits}
\title{Combine BUGS fits}
\usage{
combineBUGSFits(fit1, ...)
}
\arguments{
\item{fit1}{an \code{"fcs2Fit"} object containing a full \acronym{FCS2} model
fit, as returned from \code{\link{fcs2FitModel}} with \code{runBUGS =
TRUE}.}

\item{...}{further \code{"fcs2Fit"} objects, each resulting from a full
\acronym{FCS2} model fit with the same model and data but perhaps different
initial values or random seed. The modified \code{"fcs2Fit"} object
\code{fit} containing the combined sample.}
}
\description{
Combines multiple BUGS fits to the same model into a single fit object. This
may be useful for combining multiple \acronym{MCMC} chains that were split
between computers to reduce computational time. \code{\link{thinBUGSSamples}}
is used when necessary with a warning to thin samples so that the number per
chain is consistent across fits.
}
\seealso{
\code{\link{thinBUGSSamples}} for thinning samples to reduce
  autocorrelation
}
