% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catch.R
\name{dCatch}
\alias{dCatch}
\alias{pCatch}
\alias{Catch}
\alias{catch}
\title{Probabilities of the total catch distribution}
\usage{
dCatch(fit, x, newData, subset = 1:nrow(newData), na.action, mu, rho)
}
\arguments{
\item{fit}{an \code{"fcs2Fit"} object containing a full \acronym{FCS2} model
fit, as returned from \code{\link{fcs2FitModel}} with \code{runBUGS =
TRUE}.}

\item{x}{vector of (non-negative integer) quantiles.}

\item{newData}{a data frame with surveys as rows and variables as columns. It
should contain all variables required by \code{fit}.}

\item{subset}{an optional vector specifying a subset of surveys to calculate
the predicted total catch for.}

\item{na.action}{a function which indicates what should happen when the data
contain missing values (\code{NA}s). The default is set by the
\code{na.action} setting of \code{\link{options}} and this is usually set
to \code{\link{na.omit}}. This setting removes surveys that contain missing
data in any required variables. A vector indicating the rows that were
removed can be extracted from the returned object using
\code{\link{na.action}}. Alternatively, \code{\link{na.pass}} can be used
to ignore missing values (where possible) or \code{\link{na.fail}} can be
given to signal an error if missing values are found.}

\item{mu}{a matrix of posterior samples of the abundance component
\eqn{\mu}{mu} can optionally be given to save recalculation if already
available. This is assumed to have been calculated from
\code{\link{abundance}} using the same arguments as above.}

\item{rho}{a matrix of posterior samples of the prevalence component
\eqn{\rho}{rho} can optionally be given to save recalculation if already
available. This is assumed to have been calculated from
\code{\link{prevalence}} using the same arguments as above.}
}
\value{
\code{dCatch} gives the probability mass function and \code{pCatch}
  the probability distribution function evaluated at values \code{x} for the
  selected surveys. In each case, a matrix is returned with rows for \code{x}
  values and columns for selected surveys.
}
\description{
Density and distribution function of the predicted total catch over all
passes.
}
\seealso{
\code{\link{fcs2FitModel}} for fitting the \acronym{FCS2} model.\cr
  \code{\link{predict.fcs2Fit}} for calculating the expected total catch.\cr
  \code{\link{plotCatchPMF}} for plotting these probabilities.\cr
  \code{\link{dzinbinom}} for the zero-inflated negative binomial
  distribution, which these functions use.
}
\keyword{distribution}
