% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_results.R
\docType{data}
\name{demo_results}
\alias{demo_results}
\title{Demo result data}
\format{
A data frame.
}
\usage{
demo_results
}
\description{
A dataset containing demo results data. For testing, created using
random seed = 42
}
\examples{
\dontrun{data(demo_results)}
}
\keyword{datasets}
