% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs2SetDefaultPriors.R
\name{.fcs2SetDefaultPriors}
\alias{.fcs2SetDefaultPriors}
\title{Set Default Priors}
\usage{
.fcs2SetDefaultPriors(fit)
}
\arguments{
\item{fit}{an object of class \code{"fcs2Fit"}, usually returned by
\code{\link{fcs2FitModel}}.}
}
\value{
the list \code{fit$prior.parameters} modified to contain parameters
for every variable in the model. %$
}
\description{
Sets default prior parameters for variables missing in
\code{fit$prior.parameters}.  %$
}
\details{
The default priors are as follows:

The shape parameter \eqn{r} follows a log-Normal distribution with
\code{mean=1} and \code{sd=10}.\cr The catch probability \eqn{q} follows a
Uniform distribution between 0 and 1.\cr The constant regression terms
\code{beta.const} and \code{gamma.const} follow a Normal distribution with
\code{mean=0} and \code{sd=5}.\cr All other linear terms \code{beta.}* or
\code{gamma.}* follow a Normal distribution with \code{mean=0} and
\code{sd=1}.\cr The hyperparameters for random walk terms \code{\link{rw1}}
or \code{\link{rw2}} are specified for the scale parameters \code{sigma.}*
or \code{nu.}* which follow the \code{\link{hyperprior}} distribution with
\code{mean=0.5} and \code{sd=3}.\cr The hyperparameters for
\code{\link{spatial}} terms are specified for the precision parameters
\code{tau.}* or \code{phi.}* and follow a Gamma distribution with
\code{shape=1} and \code{rate=0.001}.

The priors should always be checked (with \code{\link{plot.fcs2Fit}} or
\code{\link{fcs2Priors}}) before fitting the full \acronym{FCS2} model with
\acronym{BUGS} and adjusted if necessary through the \code{prior.parameters}
argument of \code{\link{fcs2FitModel}}.
}
\note{
The user should not call this function directly but it is called by
\code{\link{fcs2FitModel}} when creating a complete \code{"fcs2Fit"} object.
}
\seealso{
\code{\link{fcs2FitModel}}
}
