% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbinom.R
\name{dzinbinom}
\alias{dzinbinom}
\alias{zinbinom}
\title{Zinbinom}
\usage{
dzinbinom(x, size, prob, zeroprob, nbmean)
}
\arguments{
\item{x}{x}

\item{size}{size}

\item{prob}{prob}

\item{zeroprob}{zero prob}

\item{nbmean}{nbmean}
}
\value{
zero-inflated Negative Binomial Distributions.
}
\description{
Zero-inflated Negative Binomial Distributions.
}
\examples{
\dontrun{
 data$Catch <- rzinbinom(100,
size = 1.1, zeroprob = 0.3,
nbmean = 0.3 * Data$SurveyArea
)
}
}
