% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs2Classify.R
\name{fcs2Classify}
\alias{fcs2Classify}
\title{Probabilistic WFD Classifications from EQR Samples}
\usage{
fcs2Classify(eqr, survey = 1:ncol(eqr), species, boundaries)
}
\arguments{
\item{eqr}{an object of class \code{"fcs2EQR"} containing Monte Carlo
\acronym{EQR} samples, as calculated from \code{\link{fcs2SingleEQR}},
\code{\link{fcs2JointEQR}} or \code{\link{fcs2JointAndSingleEQR}}.}

\item{survey}{index specifying which surveys (or water bodies etc if surveys
joined) to classify.}

\item{species}{index specifying which species to classify.}

\item{boundaries}{a vector of length 4 giving the \acronym{EQR} boundaries
separating the classes \emph{Bad}, \emph{Poor}, \emph{Good},
\emph{Moderate} and \emph{High}. If missing, regularly spaced boundaries of
\code{c(0.2, 0.4, 0.6, 0.8)} are used with a warning.}
}
\value{
A matrix or array containing the probabilities of each \acronym{WFD}
  class, as estimated from the proportion of \acronym{EQR} samples between
  the class boundaries. The classes \emph{Bad}, \emph{Poor}, \emph{Good},
  \emph{Moderate} and \emph{High} are given in rows 1 to 5 respectively,
  while the columns indicate the selected surveys (or sites/water bodies/etc
  if surveys were joined when calculating \acronym{EQR}s). If \code{eqr}
  contains \acronym{EQR} samples for multiple fits/species, an array is
  returned with species as the third dimension.
}
\description{
Produces probabilistic Water Framework Directive classifications of
\emph{Bad}, \emph{Poor}, \emph{Good}, \emph{Moderate} or \emph{High} from
Monte Carlo \acronym{EQR} samples.
}
\examples{
    \dontrun{
### Very simple example using a single species EQR

# simulate random dataset
Data <- data.frame(SurveyArea=rlnorm(100, 4.6, 0.5))

# random survey area
Data$Catch <- rzinbinom(100, size=1.1, zeroprob=0.3,
                  nbmean=0.3 * Data$SurveyArea)  # single catch per survey

# define a simple model with no covariates
# and fit full model with OpenBUGS
fit <- fcs2FitModel("Catch", dataFrame=Data, surveyAreaVar="SurveyArea",
                           runBUGS=TRUE, bugsProgram="OpenBUGS", n.iter=1000)

# calculate samples of single EQR, using same dataset
eqr <- fcs2SingleEQR(fit, Data)

# define WFD class boundaries
boundaries <- c(0.001, 0.01, 0.25, 0.5)

# plot EQR variables for first 9 surveys,
# shading the probabilities of each class
plot(eqr, 1:9, type="density", boundaries=boundaries)

# calculate probabilistic classifications
fcs2Classify(eqr, boundaries=boundaries)

# plot these probabilities for the first 9 surveys
plot(eqr, 1:9, type="class", boundaries=boundaries)
}
}
\seealso{
\code{\link{plot.fcs2EQR}} with \code{type="class"} for plotting the
  probabilistic classifications.#' \code{\link{fcs2SingleEQR}},
  \code{\link{fcs2JointEQR}} or \code{\link{fcs2JointAndSingleEQR}} for
  producing \code{"fcs2EQR"} objects.
}
