% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs2EQRSummaryMatrix.R
\name{fcs2EQRSummaryMatrix}
\alias{fcs2EQRSummaryMatrix}
\title{EQR Summary Matrix}
\usage{
fcs2EQRSummaryMatrix(fit1, ..., newData, joinByVar = NULL,
  subset = 1:nrow(newData), na.action, classify = !missing(boundaries),
  boundaries, summary = "mean", observations = TRUE,
  predictions = TRUE, n.samples = 1000, n.sims = 1000, eqrs,
  seed = NULL)
}
\arguments{
\item{fit1}{an \code{"fcs2Fit"} object containing a full \acronym{FCS2} model
fit, as returned from \code{\link{fcs2FitModel}} with \code{runBUGS =
TRUE}.}

\item{...}{further \code{"fcs2Fit"} objects, each resulting from a full
\acronym{FCS2} model fit with a different species.}

\item{newData}{a data frame with surveys as rows and variables as columns. It
should contain all variables required by each of the model fits. Covariates
which are related to human disturbance (\dfn{pressure variables}) should
have their values set to the value expected at the site if it were
undisturbed (\dfn{reference conditions}) rather than the observed value for
each of these variables.}

\item{joinByVar}{the name of a column in \code{dataFrame} to be used for
joining multiple surveys together when calculating the \acronym{EQR}. For
example, \code{joinByVar = "WBId"} would produce \acronym{EQR} samples for
each water body, with every survey in a water body contributing towards its
\acronym{EQR}.\cr If \code{NULL} (the default), \acronym{EQR}s are
calculated for each survey only.}

\item{subset}{an optional vector specifying a subset of surveys to calculate
\acronym{EQR} samples for.}

\item{na.action}{a function which indicates what should happen when the data
contain missing values (\code{NA}s). The default is set by the
\code{na.action} setting of \code{\link{options}} and this is usually set
to \code{\link{na.omit}}. This setting removes surveys that contain missing
data in any required variables. Alternatively, \code{\link{na.pass}} can be
used to ignore missing values (where possible) or \code{\link{na.fail}} can
be given to signal an error if missing values are found.}

\item{classify}{whether to produce probabilistic classifications of each
\acronym{WFD} class.  If \code{joinByVar = NULL} (the default),
classifications are made using the joint \acronym{EQR} for each survey.  If
\code{joinByVar} is provided, classifications are made using the combined
\acronym{EQR}s across multiple surveys.}

\item{boundaries}{a vector of length 4 giving the \acronym{EQR} boundaries
separating the classes \emph{Bad}, \emph{Poor}, \emph{Good},
\emph{Moderate} and \emph{High}. This is used if \code{classify = TRUE} to
produce probabilistic classifications of each class. If missing, regularly
spaced boundaries of \code{c(0.2, 0.4, 0.6, 0.8)} are used with a warning.}

\item{summary}{a character vector listing how to summarise the \acronym{EQR}
variables. Can contain \code{"mean"} for the mean and/or \code{"sd"} to
calculate the standard deviation.}

\item{observations}{whether to give the observed total catch for each survey
and species.}

\item{predictions}{whether to summarise the model's predicted total catch for
each survey and species. If \code{TRUE} (the default), the expected total
catch is given but if \code{"detail"}, the probability of presence
(prevalence) and the expected total catch if present are alternatively
given. If \code{"all"}, all three columns are shown for each species.}

\item{n.samples}{the number of Monte Carlo \acronym{EQR} samples to produce
for each survey (or joining variable).}

\item{n.sims}{the number of Monte Carlo simulations to make for each
\acronym{EQR} sample. These internal samples are used for approximating the
probability that defines the joint \acronym{EQR}.}

\item{eqrs}{the result of \code{\link{fcs2JointAndSingleEQR}} with the above
arguments can alternatively be given, to save recalculation if available.}

\item{seed}{set random seed to allow repeatable results.}
}
\value{
a matrix with surveys as rows and columns containing the selected
  information.
}
\description{
Produces a matrix summarising the single and joint \acronym{EQR}s and
optionally comparing these to the observed catches and the \acronym{FCS2}
model's predictions.
}
\seealso{
\code{\link{fcs2JointAndSingleEQR}}
}
