% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs2InteractiveLikelihood.R
\name{fcs2InteractiveLikelihood}
\alias{fcs2InteractiveLikelihood}
\title{Interactive FCS2 model likelihood}
\usage{
fcs2InteractiveLikelihood(multipass = FALSE, eqr = FALSE,
  boundaries = NULL, r.limit = c(0.1, 10), mu.limit = c(0, 10),
  rho.limit = c(0, 1), a.limit = c(0, 10), d.limit = c(1, 7),
  q.limit = c(0, 1), c.limit = c(0, 20))
}
\arguments{
\item{multipass}{If \code{FALSE} (the default), the \acronym{ZINB}
likelihood of the single catch \eqn{C} from the original \acronym{EA}
\acronym{FCS2} model is given.  If \code{TRUE}, the \acronym{ZINB}
distribution of the total catch \eqn{T} over all passes from the
multiple-pass \acronym{FCS2} model is given.  The true likelihood for this
model is the multivariate \acronym{ZINM} distribution (see
\code{fcs2:::zinmultinom}).}

\item{eqr}{whether to display the \acronym{EQR} value.}

\item{boundaries}{a vector of length 4 giving the \acronym{EQR} boundaries
separating the classes \emph{Bad}, \emph{Poor}, \emph{Good}, \emph{Moderate}
and \emph{High}.  These are used only to colour the plot with \emph{Bad} red
and \emph{High} blue.  If \code{NULL} (default), the probability that
defines the single \acronym{EQR} is coloured blue.}

\item{r.limit, mu.limit, rho.limit, a.limit}{vectors of giving the lower and
upper limits for the interactive controls for the shape parameter \eqn{r},
the abundance \eqn{\mu}, the prevalence \eqn{rho} and the survey area
\eqn{a} respectively.}

\item{d.limit, q.limit}{vectors of giving the lower and upper limits for the
interactive controls for the number of runs \eqn{d} and the catch
probability \eqn{q} respectively.  These are used only if \code{multipass =
TRUE}.}

\item{c.limit}{vector of giving the lower and upper limit for the
interactive control for the single catch.  This is used only if \code{eqr =
TRUE}.}
}
\description{
Creates a plot of the \acronym{FCS2} model likelihood with interactive
controls that allow the model parameters to be adjusted.  The likelihood
represents the probability of obtaining a single fish count and the controls
allow you to see how these probabilities change with the parameters.  The
function can visualise either the zero-inflated negative binomial
(\acronym{ZINB}) distribution (see \code{\link{zinbinom}}) of the single catch
\eqn{C} in the original \acronym{EA} \acronym{FCS2} model or the
\acronym{ZINB} distribution of the total catch \eqn{T} over all passes from
the multiple-pass \acronym{FCS2} model.
}
\section{Warning}{
 This function requires the additional package \pkg{rpanel}
for producing interactive controls.
}

\examples{

## Interactive likelihood of total catch with EQR given against class boundaries

# only run example if R is running interactively
if (interactive()) {
    # define WFD class boundaries
    boundaries <- c(0.001, 0.01, 0.25, 0.5)

    # create interactive plot
    fcs2InteractiveLikelihood(multipass=TRUE, eqr=TRUE, boundaries=boundaries)
}

}
\seealso{
\code{\link{fcs2InteractivePrediction}} which plot predictions from
a fitted \acronym{FCS2} model with interactive controls to show how the
predictions change when covariates are varied.
}
\keyword{hplot}
