% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs2InteractivePrediction.R
\name{fcs2InteractivePrediction}
\alias{fcs2InteractivePrediction}
\title{Interactive FCS2 model prediction}
\usage{
fcs2InteractivePrediction(fit, data, init.row, eqr = TRUE, boundaries = NULL)
}
\arguments{
\item{fit}{an \code{"fcs2Fit"} object containing a full \acronym{FCS2} model
fit, as returned from \code{\link{fcs2FitModel}} with \code{runBUGS = TRUE}.}

\item{data}{a data frame with surveys as rows and variables as columns. It
should contain all variables required by \code{fit} and is used primarily to
specify the upper and lower limits for the interactive controls for each
covariate.}

\item{init.row}{optionally, an integer corresponding to a row in \code{data}
that should be used as the initial state for each covariate.  If omitted or
the selected value is missing, the initial state for a covariate is selected
as the average value in \code{data}.}

\item{eqr}{whether to give the distribution of the \acronym{EQR} variable.
Defaults to \code{TRUE}.}

\item{boundaries}{a vector of length 4 giving the \acronym{EQR} boundaries
separating the classes \emph{Bad}, \emph{Poor}, \emph{Good}, \emph{Moderate}
and \emph{High}.  These are used only to colour the plots with \emph{Bad}
red and \emph{High} blue.  If \code{NULL} (default), the probability that
defines the single \acronym{EQR} is coloured blue.}
}
\description{
Plots the posterior distribution of the abundance and prevalence components
as well as the predictive distribution of a new fish catch at a survey
characterised by model covariates that can be modified interactively.
Controls are provided for each covariate in the model so that the change in
predictive probabilities can be visualised as covariates are adjusted.  The
single \acronym{EQR} variable can additionally be displayed and class
boundaries can be provided to colour the probabilities of each class.
}
\section{Warning}{
 This function requires the additional package \pkg{rpanel}
for producing interactive controls.
}

\seealso{
\code{\link{fcs2InteractiveLikelihood}} which produces a simpler
interactive demonstration of how the \acronym{FCS2} \acronym{ZINB}
likelihood varies with model parameters.
}
\keyword{hplot}
