% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs2JointEQR.R
\name{fcs2JointEQR}
\alias{fcs2JointEQR}
\title{Joint EQR}
\usage{
fcs2JointEQR(
  fit1,
  ...,
  newData,
  joinByVar = NULL,
  subset = 1:nrow(newData),
  na.action,
  n.samples = 1000,
  n.sims = 1000,
  both = FALSE,
  showProgress = TRUE
)
}
\arguments{
\item{fit1}{an \code{"fcs2Fit"} object containing a full \acronym{FCS2}
model fit, as returned from \code{\link{fcs2FitModel}} with \code{runBUGS =
TRUE}.}

\item{\dots}{further \code{"fcs2Fit"} objects, each resulting from a full
\acronym{FCS2} model fit with a different species.}

\item{newData}{a data frame with surveys as rows and variables as columns.
It should contain all variables required by each of the model fits.
Covariates which are related to human disturbance (\dfn{pressure variables})
should have their values set to the value expected at the site if it were
undisturbed (\dfn{reference conditions}) rather than the observed value for
each of these variables.}

\item{joinByVar}{the name of a column in \code{dataFrame} to be used for
joining multiple surveys together when calculating the \acronym{EQR}.  For
example, \code{joinByVar = "WBId"} would produce \acronym{EQR} samples for
each water body, with every survey in a water body contributing towards its
\acronym{EQR}.\cr If \code{NULL} (the default), \acronym{EQR} samples are
calculated for each survey.}

\item{subset}{an optional vector specifying a subset of surveys to calculate
\acronym{EQR} samples for.}

\item{na.action}{a function which indicates what should happen when the data
contain missing values (\code{NA}s).  The default is set by the
\code{na.action} setting of \code{\link{options}} and this is usually set to
\code{\link{na.omit}}.  This setting removes surveys that contain missing
data in any required variables.  A vector indicating the rows that were
removed can be extracted from the returned object using
\code{\link{na.action.fcs2EQR}} if surveys are not joined.  Alternatively,
\code{\link{na.pass}} can be used to ignore missing values (where possible)
or \code{\link{na.fail}} can be given to signal an error if missing values
are found.}

\item{n.samples}{the number of Monte Carlo \acronym{EQR} samples to produce
for each survey (or joining variable).}

\item{n.sims}{the number of Monte Carlo simulations to make for each
\acronym{EQR} sample.  These internal samples are used for approximating the
probability that defines the joint \acronym{EQR}.}

\item{both}{if \code{TRUE} and \code{joinByVar} is provided, \acronym{EQR}
samples are produced both for each survey and also for each level of the
joining variable.  This allows for efficient calculation of the joint
\acronym{EQR} for each survey and also the joint \acronym{EQR} that also
joins surveys.}

\item{showProgress}{whether to display the current progress while generating
\acronym{EQR} samples.}
}
\value{
If \code{both = FALSE} (the default), a single \code{"fcs2EQR"}
object containing Monte Carlo \acronym{EQR} samples is returned.  The
\code{"fcs2EQR"} object is essentially a matrix with Monte Carlo samples as
rows and columns corresponding to either surveys, if \code{joinByVar} is
missing, or to different levels of the joining variable, e.g. different
water bodies.

If \code{both = TRUE} and \code{joinByVar} is provided, a list containing
both of these \code{"fcs2EQR"} objects is returned.
}
\description{
Produces Monte Carlo samples of the joint \dfn{Ecological Quality Ratio}
(\acronym{EQR}) that combines multiple species and optionally multiple
surveys within a group (such as a water body).  This provides a single
\acronym{EQR} variable that is found by comparing the observed fish catches
with the model's predictions at reference conditions for a number of species
and optionally a number of surveys.\cr This function can be used to provide
a single \acronym{EQR} variable for each water body that can be passed to
\code{\link{fcs2Classify}} to produce probabilistic \acronym{WFD}
classifications.
}
\section{Warning}{
 Although this function has been written in C for speed, it
can still take a long time to produce a large number \code{n.samples} of
\acronym{EQR} samples when the number \code{n.sims} of simulations is also
high.
}

\examples{

\dontrun{

### Example 1: Simple example with two model fits and no covariates
###

# simulate random dataset for example
Data <- data.frame(SurveyArea=rlnorm(100, 4.6, 0.5))   # random survey area
# a single salmon catch per survey
Data$Salmon <- rzinbinom(100, size=1.1,
                         zeroprob=0.3, nbmean=0.3 * Data$SurveyArea)
# a single trout catch per survey
Data$Trout <- rzinbinom(100, size=0.87,
                        zeroprob=0.19, nbmean=0.42 * Data$SurveyArea)

# fit full model for salmon with OpenBUGS
salmonFit <- fcs2FitModel("Salmon", dataFrame=Data, surveyAreaVar="SurveyArea",
                          runBUGS=TRUE, n.iter=1000, bugsProgram="OpenBUGS")

# fit full model for trout with OpenBUGS
troutFit <- fcs2FitModel("Trout", dataFrame=Data, surveyAreaVar="SurveyArea",
                         runBUGS=TRUE, n.iter=1000, bugsProgram="OpenBUGS")

# calculate samples of EQR that combines salmon and trout for each survey,
#   using same dataset as no pressure variables to adjust to reference values
eqr <- fcs2JointEQR(salmonFit, troutFit, newData=Data,
                    n.samples=100, n.sims=100)

# plot EQR variables for first 9 surveys
plot(eqr, 1:9, boundaries=NULL)

# calculate mean EQR values
mean(eqr)



### Example 2: Joining surveys as well as species to create water body EQRs
###

# extend dataset to include water body indicator
# randomly assign water body A, B or C to each survey
Data$WaterBody <- sample(c("A", "B", "C"), 100, replace=TRUE)

# calculate samples of EQR that combines salmon and trout
#   as well as combining all surveys within each water body
eqr <- fcs2JointEQR(salmonFit, troutFit, newData=Data,
                    n.samples=100, n.sims=100, joinByVar="WaterBody")

# plot EQR variables for all 3 water bodies
plot(eqr, 1:3, boundaries=NULL)

# calculate mean EQR values
mean(eqr)
}

}
\seealso{
\code{\link{print.fcs2EQR}}, \code{\link{summary.fcs2EQR}} and
\code{\link{fcs2EQRSummaryMatrix}} for summarising \code{"fcs2EQR"}
objects;\cr \code{\link{plot.fcs2EQR}} for plotting \acronym{EQR}
variables;\cr \code{\link{mean.fcs2EQR}} and \code{\link{quantile.fcs2EQR}}
for calculating means and quantiles of \acronym{EQR} variables
respectively;\cr

\code{\link{fcs2FitModel}} for producing the required \acronym{FCS2} model
fits;\cr \code{\link{fcs2Classify}} for using \acronym{EQR} samples to
produce probabilistic \acronym{WFD} classifications.

\code{\link{fcs2SingleEQR}} for producing \acronym{EQR} samples for a single
species and survey;\cr \code{\link{fcs2JointAndSingleEQR}} for calculating
joint and single \acronym{EQR} samples simultaneously.
}
