% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs2ModelSelection.R
\name{fcs2ModelSelection}
\alias{fcs2ModelSelection}
\title{Automated Model Selection}
\usage{
fcs2ModelSelection(
  runTotalVars = NULL,
  allRunsTotalVar = NULL,
  allRunsRangeVars = NULL,
  dataFrame,
  surveyAreaVar = "SurveyArea",
  nRunsVar = NULL,
  muVars,
  muVarType,
  rhoVars = muVars,
  rhoVarType = muVarType,
  rhoFormula,
  subset = 1:nrow(dataFrame),
  tolerance = 0.01,
  maxNoOrders = 3,
  nSweeps = 1,
  prior.parameters = list(),
  estAllRunsTotalVar = NULL,
  verbose = FALSE
)
}
\arguments{
\item{runTotalVars}{a character vector of columns in \code{dataFrame} that
give the number of fish caught in each run of a multiple-run survey.  These
should be in order from the first run to the last.}

\item{allRunsTotalVar}{the name of a column in \code{dataFrame} that gives
the total number of fish caught over all runs of a multiple-run survey.}

\item{allRunsRangeVars}{the names of two columns in \code{dataFrame} that
give the minimum and the maximum value of a range of possible values for the
total number of fish caught over all runs in a multiple-run survey.}

\item{dataFrame}{a data frame with surveys as rows and variables as columns.
It should contain all variables specified by other arguments.}

\item{surveyAreaVar}{the name of a column in \code{dataFrame} that gives the
survey area.  If not specified, the function will search for the default
value \code{"SurveyArea"}.}

\item{nRunsVar}{the name of a column in \code{dataFrame} that gives the
number of runs in each survey.  If missing, the number of runs is assumed to
be the number of non-missing run total entries, unless \code{runTotalVars}
has length 1 or is missing in which case a single-run model is used.  Number
of runs values greater than the number of run total columns are clipped with
a warning.}

\item{muVars}{a character vector naming variables to use for terms in the
abundance regression equation.  Variables are attempted one-by-one in the
order given.}

\item{muVarType}{a character vector of the same length as \code{muVars}
indicating the type of abundance term to attempt for each variable.  Each
element should be one of \code{"asis"}, \code{"factor"}, \code{"linear"},
\code{"continuous"} or \code{"spatial"}.  See \sQuote{Details} for a
description of each.}

\item{rhoVars}{a character vector naming variables to use for terms in the
prevalence regression equation.  Variables are attempted one-by-one in the
order given.  Defaults to \code{muVars} to use the same variables as
specified for abundance.}

\item{rhoVarType}{a character vector of the same length as \code{muVars}
indicating the type of prevalence term to attempt for each variable.  Each
element should be one of \code{"asis"}, \code{"factor"}, \code{"linear"},
\code{"continuous"} or \code{"spatial"}.  See \sQuote{Details} for a
description of each.  Defaults to \code{muVarType} to use the same variables
as specified for abundance.}

\item{rhoFormula}{an optional \code{\link{formula}} specifying which terms
should appear in the prevalence regression equation when selecting terms for
abundance.  If specified, the prevalence model selection is skipped.}

\item{subset}{an optional vector specifying a subset of surveys to be used
in the fitting process.}

\item{tolerance}{the threshold for each term's significance probability,
below which a term is considered to be significant and is retained.  The
default value is \code{0.01} but a smaller value will cause fewer terms to
be retained and vice-versa.\cr See \code{\link{summary.fcs2Fit}} for a
definition of the significance probability for each variable.  The
probabilities corresponding to each variable within a \code{\link{rw2}} or
\code{\link{spatial}} term are combined (by rescaling the minimum under the
assumption of independence) before comparison with \code{tolerance}.}

\item{maxNoOrders}{the maximum number of polynomial orders to try for each
linear component.  Defaults to 3.}

\item{nSweeps}{the number of sweeps to make through each list of potential
variables.  The default is \code{2} so that a second test is made to each
term in case the presence of later variables will make an earlier term
significant.}

\item{prior.parameters}{an optional named list of named vectors giving the
parameter values to use for the prior distribution of a variable.  See
\code{\link{fcs2Priors}} for further details on the prior distributions and
how to specify prior parameters.  Priors that are not given parameter values
have defaults given to them by \code{\link{.fcs2SetDefaultPriors}}.  Note
that priors for linear variables are ignored by \acronym{INLA}.}

\item{estAllRunsTotalVar}{the name of a column in \code{dataFrame} that
gives an estimate of the total number of fish caught over all runs for
surveys where only range data is available.  This is used in the approximate
abundance \acronym{INLA} fit only and is not used in the full model as
fitted with \acronym{BUGS}.  If this is not provided and range data are
present, the central value of each range is used with a warning.}

\item{verbose}{whether to print progress to screen.}
}
\value{
a list of two matricies, each containing a summary of the terms
attempted in each iteration.  The first matrix gives the term history for
the prevalence (\eqn{\rho}{rho}) and the second gives the history for the
abundance (\eqn{\mu}{mu}).

Each row of a matrix summarises the regression formula used for that
component, with covariates appearing as columns.  Linear terms are
represented as a number giving the order of the term, random walk terms are
given by \code{"rw2"} and spatial terms by \code{"spatial"}.
\code{fcs2:::termSummary2Formula} can be used to convert a row to an formula
for simplier input into \code{\link{fcs2FitModel}}.
}
\description{
Automatically selects an optimal set of covariate terms for the abundance
and prevalence regression equations of the \acronym{FCS2} model.  Terms are
attempted sequentially and the approximate abundance and prevalence
\acronym{INLA} fits are used to test the significance of each term.
}
\note{
Since the prevalence terms are selected first and then the abundance
terms second, it is possible that some prevalence terms are no longer
significant in the final fit as these were not attempted with the selected
abundance formula.
}
\seealso{
\code{\link{fcs2FitModel}}, \code{fcs2:::termSummary2Formula}
}
