% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs2Priors.R
\name{fcs2Priors}
\alias{fcs2Priors}
\title{Prior Distributions in the FCS2 Model}
\usage{
fcs2Priors(parameters = list(), print = TRUE, plot = TRUE, prob = 0.95, ...)
}
\arguments{
\item{parameters}{a named list of named vectors, each specifying the
parameter values for a single prior distribution.  See \sQuote{Details} for
how these are specified.  Alternatively, an \code{"fcs2Fit"} object can be
provided, as returned from \code{\link{fcs2FitModel}}.}

\item{print}{whether to print a summary of each prior distribution.}

\item{plot}{whether to plot each prior distribution.}

\item{prob}{a decimal between 0 and 1 giving the size of the confidence
interval displayed for each prior (if \code{print = TRUE}).}

\item{\dots}{further arguments passed to \code{\link{plot}} if plotting each
prior.}
}
\value{
Invisibly returns a modified list of prior parameters with each
prior specified using standardised parameters, as required for
\acronym{INLA} and \acronym{BUGS}:

The log-Normal prior for the shape parameter \eqn{r} is given by the mean
\code{mu} and precision \code{tau} of the equivalent Normal prior for
\eqn{log(r)}.\cr The Beta prior for the catch probability \eqn{q} is given
by two shape parameters \code{a} and \code{b}.\cr The Normal prior for a
linear variable \code{beta.}* or \code{gamma.}* is given by the \code{mean}
and \code{precision}.\cr The Gamma prior for a precision hyperparameter
\code{tau.}* or \code{phi.}* is given by the shape \code{a} and rate
\code{b}.

Prior parameters specified for a scale hyperparameter \code{sigma.}* or
\code{nu.}* (following the \code{\link{hyperprior}} distribution) are
transformed to the corresponding precision parameter \code{tau.}* or
\code{phi.}* respectively.
}
\description{
Summarises and plots the prior distribution for each parameter in the
\acronym{FCS2} model.  This can be used to aid the selection of parameters
for each prior.
}
\details{
This function can also be used to translate a user-friendly specification of
each prior distribution to the form required by \acronym{INLA} and
\acronym{BUGS} to fit the model.

Each prior distribution can be specified either by their \code{mean} and
standard deviation \code{sd}, \code{mean} and variance \code{var} or by
providing values for their parameters.
}
\examples{

# Specify a range of priors by their parameters:
# each prior is plotted and summarised by default
fcs2Priors(list(r=c(mu=0, sigma=3),
                q=c(a=0.5, b=0.5),
                beta=c(mean=0, precision=0.5),
                phi=c(a=1, b=0.3)))

# Priors can also be specified by their mean and sd or var:
fcs2Priors(list(r=c(mean=3, sd=10),
                q=c(mean=0.6, sd=0.2),
                beta=c(mean=0, var=2),
                sigma=c(mean=1.5, sd=3)))

}
\seealso{
\code{\link{fcs2FitModel}}
}
