% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs2RunINLA.R
\name{fcs2RunINLA}
\alias{fcs2RunINLA}
\title{Fit Approximate FCS2 Models with INLA}
\usage{
fcs2RunINLA(fit, run = TRUE, verbose = TRUE)
}
\arguments{
\item{fit}{an object of class \code{"fcs2Fit"}, usually returned by
\code{\link{fcs2FitModel}}.}

\item{run}{which approximate \acronym{FCS2} models to fit with
\acronym{INLA}.  If \code{TRUE} (the default) or \code{"both"},
\acronym{INLA} is run for both models.  Alternatively, if \code{"rho"} or
\code{"prevalence"} the prevalence model only is run and if \code{"mu"} or
\code{"abundance"} the approximate abundance model only is run.  The
abundance model can only be run if the prevalence model is either run first
or is provided through \code{fit}.}

\item{verbose}{whether to print progress to screen.}
}
\value{
a list with two components: \item{rhoFit}{ an \code{"inla"} object
containing the approximate prevalence model fit } \item{muFit}{ an
\code{"inla"} object containing the approximate abundance model fit, or
\code{NULL} if this was not calculated or provided } See \code{inla}
from the package \pkg{INLA} for a description of the \code{"inla"} object.
}
\description{
Fits two approximate \acronym{FCS2} sub-models using \dfn{Integrated Nested
Laplace Approximations} (\acronym{INLA}).
}
\details{
\acronym{INLA} is used to fit two models that together approximate the full
\acronym{FCS2} statistical model (see \code{\link{fcs2FitModel}}.

Firstly, an approximate presence/absence model is fitted to estimate the
prevalence regression.  This assumes the observed presence follows a
Bernoulli distribution with probability of presence \eqn{\rho} (the
prevalence).  The \code{\link{logit}} regression for \eqn{\rho} takes the
same form as in the full \acronym{FCS2} model.

After fitting the prevalence sub-model, the presence of the species can be
estimated.  Data for surveys either observed or predicted to be present are
then modelled by the second model that attempts to estimate the abundance
regression.  This models the total catch \eqn{T} over all passes by the
Negative Binomial distribution with shape \eqn{r} and mean given by \eqn{a
\mu}, where \eqn{a} is the survey area and \eqn{\mu} is the abundance.  The
regression for \eqn{\mu} takes the same form as in the full \acronym{FCS2}
model.

This model is exact for the original \acronym{FCS2} model but only an
approximation of the total catch when present in the multiple-pass model
since it ignores the dependence upon the catch probability \eqn{q} and the
number of runs \eqn{d}. However, since \eqn{q} is constant and if the number
of runs does not vary significantly, this discrepancy should be mostly
absorbed into an error in the constant parameter \eqn{\beta_0}.  If this is
true, the model estimates can still be used to estimate the regression terms
so that their significance can be judged.
}
\note{
The user would not usually call this function directly as it is called
if requested by \code{\link{fcs2FitModel}}.
}
\seealso{
\code{\link{fcs2FitModel}} for the preferred way of calling this
function. \cr \code{inla} from the package \pkg{INLA} for details of
the \acronym{INLA} estimation method and the \code{"inla"} objects it
returns.
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{spatial}
