% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs2SingleEQR.R
\name{fcs2SingleEQR}
\alias{fcs2SingleEQR}
\title{Single EQR}
\usage{
fcs2SingleEQR(fit, newData, subset = 1:nrow(newData), na.action, mu, rho)
}
\arguments{
\item{fit}{an \code{"fcs2Fit"} object containing a full \acronym{FCS2} model
fit, as returned from \code{\link{fcs2FitModel}} with \code{runBUGS = TRUE}.}

\item{newData}{a data frame with surveys as rows and variables as columns.
It should contain all variables required by \code{fit}.  Covariates which
are related to human disturbance (\dfn{pressure variables}) should have
their values set to the value expected at the site if it were undisturbed
(\dfn{reference conditions}) rather than the observed value for each of
these variables.}

\item{subset}{an optional vector specifying a subset of surveys to calculate
\acronym{EQR} samples for.}

\item{na.action}{a function which indicates what should happen when the data
contain missing values (\code{NA}s).  The default is set by the
\code{na.action} setting of \code{\link{options}} and this is usually set to
\code{\link{na.omit}}.  This setting removes surveys that contain missing
data in any required variables.  A vector indicating the rows that were
removed can be extracted from the returned object using
\code{\link{na.action.fcs2EQR}}.  Alternatively, \code{\link{na.pass}} can
be used to ignore missing values (where possible) or \code{\link{na.fail}}
can be given to signal an error if missing values are found.}

\item{mu}{a matrix of posterior samples of the abundance component \eqn{\mu}
can optionally be given to save recalculation if already available.  This is
assumed to have been calculated from \code{\link{abundance}} using the same
arguments as above.}

\item{rho}{a matrix of posterior samples of the prevalence component
\eqn{\rho} can optionally be given to save recalculation if already
available.  This is assumed to have been calculated from
\code{\link{prevalence}} using the same arguments as above.}
}
\value{
Returns an \code{"fcs2EQR"} object that contains the \acronym{EQR}
samples.  The \code{"fcs2EQR"} object is essentially a matrix with Monte
Carlo samples as rows and surveys as columns.
}
\description{
Produces Monte Carlo samples of the \acronym{FCS2} \dfn{Ecological Quality
Ratio} (\acronym{EQR}) for a single species and for each survey.  The
\acronym{EQR} is a number between 0 and 1 which is found by comparing the
observed catch with the model's prediction at reference conditions.  Higher
\acronym{EQR} values are better and these are caused by catching larger
numbers of fish.  The variability in the Monte Carlo samples of the
\acronym{EQR} is due to the uncertainty in the parameters of the
\acronym{FCS2} model.  This uncertainty can be used by
\code{\link{fcs2Classify}} to produce probabilistic \acronym{WFD}
classifications from each \acronym{EQR}.
}
\examples{

\dontrun{

### Very simple example with no covariates
###

# simulate random dataset
Data <- data.frame(SurveyArea=rlnorm(100, 4.6, 0.5))   # random survey area
Data$Catch <- rzinbinom(100, size=1.1, zeroprob=0.3,
                        nbmean=0.3 * Data$SurveyArea)  # single catch per survey

# fit full model with OpenBUGS
fit <- fcs2FitModel("Catch", dataFrame=Data, surveyAreaVar="SurveyArea",
                    runBUGS=TRUE, n.iter=1000, bugsProgram="OpenBUGS")

# calculate samples of single EQR, using same dataset
eqr <- fcs2SingleEQR(fit, Data)

# plot EQR variables for first 9 surveys
plot(eqr, 1:9, boundaries=NULL)

# calculate mean EQR values
mean(eqr)
}

}
\seealso{
\code{\link{print.fcs2EQR}}, \code{\link{summary.fcs2EQR}} and
\code{\link{fcs2EQRSummaryMatrix}} for summarising \code{"fcs2EQR"}
objects;\cr \code{\link{plot.fcs2EQR}} for plotting \acronym{EQR}
variables;\cr \code{\link{mean.fcs2EQR}} and \code{\link{quantile.fcs2EQR}}
for calculating means and quantiles of \acronym{EQR} variables
respectively;\cr

\code{\link{fcs2FitModel}} for producing the required \acronym{FCS2} model
fit;\cr \code{\link{fcs2Classify}} for using \acronym{EQR} samples to
produce probabilistic \acronym{WFD} classifications.

\code{\link{fcs2JointEQR}} for producing \acronym{EQR} samples that combine
multiple species and/or surveys;\cr \code{\link{fcs2JointAndSingleEQR}} for
calculating joint and single \acronym{EQR} samples simultaneously.
}
