% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs2SpatialSummary.R
\name{fcs2SpatialSummary}
\alias{fcs2SpatialSummary}
\title{Summary of Spatial Terms in FCS2 Model}
\usage{
fcs2SpatialSummary(
  map,
  fit,
  posterior = !is.null(fit$bugsFit),
  inla = !posterior,
  abundance = TRUE,
  prevalence = TRUE
)
}
\arguments{
\item{map}{a \code{sp::SpatialPolygonsDataFrame} object containing
polygons and a data frame corresponding to the regions used for a spatial
term in the \acronym{FCS2} model.  This will usually be generated from an
ESRI shapefile using \code{readShapePoly}.}

\item{fit}{an \code{"fcs2Fit"} object containing \acronym{INLA} or
\acronym{BUGS} model fits, as returned from \code{\link{fcs2FitModel}}.  The
model should include a \code{\link{spatial}} term in the abundance and/or
prevalence component which is based upon the region given by \code{map}.}

\item{posterior}{whether to summarise the posterior estimates of each region
(if available).}

\item{inla}{whether to summarise the approximate \acronym{INLA} estimates of
each region.}

\item{abundance}{whether the spatial term (with these regions) is in the
abundance regression component.}

\item{prevalence}{whether the spatial term (with these regions) is in the
prevalence regression component.}
}
\value{
the \code{sp::SpatialPolygonsDataFrame} object \code{map}
with additional columns in the data frame component \code{map@data}
containing the requested summaries.  Each summary is given as two columns,
one for the mean of each variable named \code{mean.}* and another for the
standard deviation named \code{sd.}*, where * is the name of the spatial
variable.
}
\description{
Summarises the spatial terms in a fitted \acronym{FCS2} model and attaches
the summary to a map of the spatial region used.
}
\section{Warning}{
 This function requires the additional package
\pkg{maptools} which may also be useful for creating \code{map} from an ESRI
shapefile using the function \code{maptools::readShapePoly}.
}

\seealso{
\code{\link{plotSpatialTerm}} which uses this function to plot the
spatial term summaries on a map.
}
