% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs2WriteModel.R
\name{fcs2WriteModel}
\alias{fcs2WriteModel}
\title{Write BUGS Model File}
\usage{
fcs2WriteModel(filename = "model.bug", muLinearVars = c(),
  muRWVars = c(), muSpatialVar = NULL, rhoLinearVars = c(),
  rhoRWVars = c(), rhoSpatialVar = NULL, multiRun = FALSE,
  catchTotal = TRUE, catchRange = FALSE, qUnif = FALSE)
}
\arguments{
\item{filename}{the name or the \acronym{BUGS} model file to create.}

\item{muLinearVars}{a character vector of linear terms in the abundance
regression}

\item{muRWVars}{a character vector of variables with first or second-order
random walk terms in the abundance regression.}

\item{muSpatialVar}{the name of the abundance spatial variable, if present.}

\item{rhoLinearVars}{a character vector of linear terms in the abundance
regression}

\item{rhoRWVars}{a character vector of variables with first or second-order
random walk terms in the prevalence regression.}

\item{rhoSpatialVar}{the name of the prevalence spatial variable, if
present.}

\item{multiRun}{whether the multiple-run extension to the \acronym{FCS2}
model is required.}

\item{catchTotal}{whether catch total data are present in the model.}

\item{catchRange}{whether catch total range data are present in the model.}

\item{qUnif}{whether to use a uniform prior for the catch probability
\eqn{q} rather than the default beta prior.  Although the uniform is a
special case of the beta, this option may help \acronym{BUGS} fit the model.}
}
\description{
Creates a \acronym{BUGS} model file that describes the structure of the full
\acronym{FCS2} statistical model.  This can be used by WinBUGS or OpenBUGS
to fit the model.
}
\note{
The user would not usually call this function directly as it is called
if required by \code{\link{fcs2FitModel}}.  However, it may be useful for
creating a \acronym{BUGS} model file to be run by WinBUGS or OpenBUGS
extenally.
}
\seealso{
\code{\link{fcs2FitModel}}
}
