% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperprior.R
\encoding{UTF-8}
\name{hyperprior}
\alias{hyperprior}
\alias{dhyperprior}
\alias{phyperprior}
\alias{qhyperprior}
\alias{rhyperprior}
\title{The Hyperprior Distribution}
\usage{
dhyperprior(x, a, b)

phyperprior(q, a, b, lower.tail = TRUE)

qhyperprior(p, a, b)

rhyperprior(n, a, b)

phyperprior(q, a, b, lower.tail = TRUE)

qhyperprior(p, a, b)

rhyperprior(n, a, b)
}
\arguments{
\item{x}{vector of quantiles.}

\item{a}{a shape parameter, equal to the shape parameter of the related Gamma
distribution.}

\item{b}{a scale parameter, equal to the rate parameter of the related Gamma
distribution.}

\item{q}{vector of quantiles.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are
\eqn{P\{X \le x\}}{P{X \le x}}, otherwise, \eqn{P\{X > x\}}{P{X > x}}}

\item{p}{vector of probabilities.}

\item{n}{number of samples required.}
}
\value{
\code{dhyperprior} gives the density, \code{phyperprior} gives the
  distribution function, \code{qhyperprior} gives the quantile function, and
  \code{rhyperprior} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for
the distribution applied as a prior for the scale hyperparameters in the
\acronym{FCS2} model.\cr This is the distribution of \eqn{\sigma = 1 /
\sqrt{\tau}} where \eqn{\tau} follows a Gamma distribution.
}
\details{
The Hyperprior Distribution

The hyperprior distribution with shape parameter \eqn{a} and scale
  parameter \eqn{b} has density \deqn{p(x) = 2 \frac{b^a}{\Gamma(a)}
  \frac{1}{x^{2a + 1}} \exp\left( -\frac{b}{x^2} \right) }{% p(x) = 2 b^a /
  (\Gamma(a) x^(2a + 1)) exp( -b / (x^2) ) } for \eqn{x, a, b > 0}.

  This is the distribution of \eqn{\sigma = 1 / \sqrt{\tau}}{\sigma = 1 /
  \sqrt(\tau)} where \eqn{\tau} follows a Gamma distribution with shape
  parameter \eqn{a} and rate parameter \eqn{b}.

  The mean is given by \eqn{\Gamma(a - 0.5) \sqrt{b} / \Gamma(a)}{\Gamma(a -
  0.5) \sqrt(b) / \Gamma(a)} if \eqn{a > 0.5} and is infinite otherwise.\cr
  The variance is given by \eqn{b / (a - 1) - b \Gamma(a - 0.5)^2 /
  \Gamma(a)^2}{b / (a - 1) - b \Gamma(a - 0.5)^2 / (\Gamma(a)^2)} if \eqn{a >
  1} and is infinite otherwise.
}
\seealso{
\code{\link{dgamma}} for the Gamma distribution.\cr
  \code{\link{specialTerms}} for the non-linear \acronym{FCS2} model terms
  that use this distribution as a prior for their scale parameters.
}
\keyword{distribution}
