% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\alias{expit}
\title{The Logit Function}
\usage{
logit(x)

expit(x)

expit(x)
}
\arguments{
\item{x}{a numeric vector or matrix}
}
\value{
A vector or matrix the same size as \code{x} containing the
transformed values.
}
\description{
The \code{logit} function and its inverse \code{expit}.  This is used as the
link function for the prevalence regression in the \acronym{FCS2} model to
transform the interval \eqn{(0, 1)} to the real line.
}
\details{
The logit transformation is defined by \eqn{logit(x) = log(x) - log(1 - x)}
so that its inverse expit is given by \eqn{expit(x) = 1 / (1 + exp(-x))}.
}
\seealso{
\code{\link{log}} for the natural logarithm and
\code{\link{prevalence}} for producing samples of the prevalence model
component, using the \code{logit} transformation.
}
