% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.fcs2EQR.R
\name{mean.fcs2EQR}
\alias{mean.fcs2EQR}
\title{Mean EQR}
\usage{
\method{mean}{fcs2EQR}(x)
}
\arguments{
\item{x}{an object of class \code{"fcs2EQR"} containing Monte Carlo
\acronym{EQR} samples, as calculated from \code{\link{fcs2SingleEQR}},
\code{\link{fcs2JointEQR}} or \code{\link{fcs2JointAndSingleEQR}}.}
}
\value{
Returns a vector or matrix containing the expected \acronym{EQR} for
each of the surveys (or sites/etc if surveys were joined) and species
contained in \code{x}.
}
\description{
Estimates the expected \acronym{EQR} from a collection of Monte Carlo
\acronym{EQR} samples.
}
\seealso{
\code{\link{quantile.fcs2EQR}}
}
