% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.action.fcs2EQR.R
\encoding{UTF-8}
\name{na.action.fcs2EQR}
\alias{na.action.fcs2EQR}
\title{Missing Values in EQR Samples}
\usage{
\method{na.action}{fcs2EQR}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"fcs2EQR"} containing Monte Carlo
\acronym{EQR} samples, as calculated from \code{\link{fcs2SingleEQR}},
\code{\link{fcs2JointEQR}} or \code{\link{fcs2JointAndSingleEQR}}.}

\item{\dots}{Not currently used.}
}
\value{
If the \acronym{EQR} object contains \acronym{EQR} samples for
multiple surveys, the returned vector indicates surveys in the data frame
that were removed due to missing observations or covariates.

Alternatively, if the \acronym{EQR} object contains \acronym{EQR} samples
corresponding to another variable (the \code{joinByVar} in the call to
\code{\link{fcs2JointEQR}} or \code{\link{fcs2JointAndSingleEQR}}, e.g.
water body ID), the returned vector indicates values of the variable that
were removed from the \acronym{EQR} object due to missing values.
}
\description{
Extract information about surveys (or water bodies, sites, etc) that were
removed when calculating the \acronym{EQR} due to missing values.
}
\details{
Missing Values in EQR Samples
}
\seealso{
\code{\link{fcs2SingleEQR}}, \code{\link{fcs2JointEQR}} and
\code{\link{fcs2JointAndSingleEQR}} for generating \acronym{EQR} samples.
}
