% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.action.fcs2Fit.R
\name{na.action.fcs2Fit}
\alias{na.action.fcs2Fit}
\title{Missing Values in FCS2 Model Fit}
\usage{
\method{na.action}{fcs2Fit}(object)
}
\arguments{
\item{object}{an object of class \code{"fcs2Fit"}, usually returned by
\code{\link{fcs2FitModel}}.}
}
\value{
the returned vector indicates surveys in the data frame that were
removed due to missing observations or covariates.
}
\description{
Extract information about surveys that were removed when fitting the
\acronym{FCS2} statistical model due to missing values.
}
\seealso{
\code{\link{fcs2FitModel}} for fitting the \acronym{FCS2} model.
}
