% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fcs2EQR.R
\name{plot.fcs2EQR}
\alias{plot.fcs2EQR}
\title{Plot EQR Samples}
\usage{
\method{plot}{fcs2EQR}(x, survey = 1, species = 1, boundaries,
  type = "density", title, adjust = 1, ...)
}
\arguments{
\item{x}{an object of class \code{"fcs2EQR"} containing Monte Carlo
\acronym{EQR} samples, as calculated from \code{\link{fcs2SingleEQR}},
\code{\link{fcs2JointEQR}} or \code{\link{fcs2JointAndSingleEQR}}.}

\item{survey}{index specifying which surveys (or water bodies etc if surveys
joined) to plot.}

\item{species}{index specifying which species to plot. This should
correspond to the plots defined by \code{survey}.}

\item{boundaries}{a vector of length 4 giving the \acronym{EQR} boundaries
separating the classes \emph{Bad}, \emph{Poor}, \emph{Good}, \emph{Moderate}
and \emph{High}.  These are used only to colour the plot.  If missing,
regularly spaced boundaries of \code{c(0.2, 0.4, 0.6, 0.8)} are used with a
warning.  If \code{NULL} and \code{type = "density"}, the plot is not
coloured.}

\item{type}{either \code{"density"} or \code{1} for a \code{\link{density}}
estimate of the \acronym{EQR} variable, or \code{"class"} or \code{2} to
plot the probability of each class (as produced by
\code{\link{fcs2Classify}}).}

\item{title}{an optional character vector giving the title for each plot.
Should have one entry per selection in \code{survey} and \code{species}.}

\item{adjust}{the bandwidth adjustment for the density estimate of the
\acronym{EQR}, if \code{type = "density"}.  See \code{\link{density}} for
further details.}

\item{\dots}{further arguments that are passed to either \code{\link{plot}}
(if \code{type = "density"}) or \code{\link{barplot}} otherwise.}
}
\description{
Plots a series of \acronym{EQR} variables or the corresponding probabilities
of each \acronym{WFD} class.
}
\seealso{
\code{\link{fcs2SingleEQR}}, \code{\link{fcs2JointEQR}} or
\code{\link{fcs2JointAndSingleEQR}} for producing \code{"fcs2EQR"}
objects.\cr

\code{\link{fcs2InteractivePrediction}} can be used to interactively view
how the \acronym{EQR} changes with covariates.
}
\keyword{hplot}
