% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fcs2Fit.R
\name{plot.fcs2Fit}
\alias{plot.fcs2Fit}
\title{Plot FCS2 Model Fit}
\usage{
\method{plot}{fcs2Fit}(x, variables = variable.names(x), prior = TRUE,
  posterior = !is.null(x$bugsFit), inla = !posterior, legend = TRUE,
  samples = FALSE, groupLinearVars = posterior)
}
\arguments{
\item{x}{an \code{"fcs2Fit"} object, as returned from
\code{\link{fcs2FitModel}}.}

\item{variables}{an optional character vector giving the names of the model
variables to plot.  \code{fcs2:::variable.names.fcs2Fit} can be used for
this.}

\item{prior}{whether to plot the prior distribution. These are only
available for plots of the density of a single variable.}

\item{posterior}{whether to plot the posterior distribution, as estimated
from \acronym{BUGS} samples (if available).}

\item{inla}{whether to plot the approximate posterior estimates produced by
\acronym{INLA} (if available).}

\item{legend}{whether to add a legend to each plot.}

\item{samples}{whether to display the individual posterior samples rather
than summarising each posterior term.}

\item{groupLinearVars}{whether to group multiple linear terms of the same
covariate together to display the overall covariate relationship rather than
each density estimate.  This is not available for the prior and
\acronym{INLA} only produces a mean estimate of the relationship based on
the false assumption of independence between the linear parameters.
Therefore the \acronym{INLA} relationship line should only be used as a
rough guide when multiple terms exist.}
}
\description{
Produces density plots of the prior and posterior estimates of each variable
in the \acronym{FCS2} model.  Can also produce series plots that display the
relationship between each covariate and the corresponding abundance or
prevalence component.
}
\seealso{
\code{\link{fcs2FitModel}}
}
\keyword{hplot}
