% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBUGSTrace.R
\name{plotBUGSTrace}
\alias{plotBUGSTrace}
\title{Trace Plots of MCMC Chains from BUGS}
\usage{
plotBUGSTrace(
  fit,
  variables = variable.names(fit, rw = "singular", spatial = "singular", hyperpar =
    "prec"),
  chains = 1:fit$bugsFit$n.chains,
  nPerPage = 12,
  col = chains + 1,
  ylab,
  ylim,
  main,
  ...
)
}
\arguments{
\item{fit}{an \code{"fcs2Fit"} object containing a full \acronym{FCS2} model
fit, as returned from \code{\link{fcs2FitModel}} with \code{runBUGS = TRUE}.}

\item{variables}{an optional character vector giving the names of the model
variables to plot.  \code{fcs2:::variable.names.fcs2Fit} can be used for
this, but only singular variable names can be given.}

\item{chains}{which \acronym{MCMC} chains to plot. Default to all.}

\item{nPerPage}{the number of trace plots to create on each page.}

\item{col}{a vector giving the colour to use for each chain in
\code{chains}.}

\item{ylab, ylim, main, ...}{further arguments passed to \code{\link{plot}}
when creating each plot.}
}
\description{
Produces trace plots of the \acronym{MCMC} samples for each parameter in the
fitted \acronym{FCS2} model.  These can be used to visualise the convergence
of the \acronym{MCMC} chains and the autocorrelation.
}
\details{
If the \acronym{MCMC} chains have converged, the multiple chains (in
different colours) should agree; Their values should not be identical but
they should cover the same range of values as each other.

The \acronym{MCMC} chains should also mix well meaning that they should move
about the parameter space quickly.  A chain that moves slowly across the
parameter space has high autocorrelation.  This can also be verified by
plotting the autocorrelation of a parameter using \code{\link{acf}}.  The
autocorrelation can be reduced by thinning the sample using
\code{\link{thinBUGSSamples}} or by specifying a larger number for
\code{n.thin} when fitting.
}
\seealso{
\code{\link{thinBUGSSamples}}
}
\keyword{hplot}
