% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCatchPMF.R
\name{plotCatchPMF}
\alias{plotCatchPMF}
\title{Plot Probability Mass Function of Predicted Total Catch}
\usage{
plotCatchPMF(fit, newData, subset = 1, na.action, boundaries, mu, rho,
  xmax, xmaxscale = 2, maxpts = 200, title, ...)
}
\arguments{
\item{fit}{an \code{"fcs2Fit"} object containing a full \acronym{FCS2} model
fit, as returned from \code{\link{fcs2FitModel}} with \code{runBUGS = TRUE}.}

\item{newData}{a data frame with surveys as rows and variables as columns.
It should contain all variables required by \code{fit}.}

\item{subset}{an vector specifying which surveys to plot the predicted
\acronym{PMF} for.}

\item{na.action}{a function which indicates what should happen when the data
contain missing values (\code{NA}s).  The default is set by the
\code{na.action} setting of \code{\link{options}} and this is usually set to
\code{\link{na.omit}}.  This setting removes surveys that contain missing
data in any required variables.  Alternatively, \code{\link{na.pass}} can be
used to ignore missing values (where possible) or \code{\link{na.fail}} can
be given to signal an error if missing values are found.}

\item{boundaries}{a vector of length 4 giving the \acronym{EQR} boundaries
separating the classes \emph{Bad}, \emph{Poor}, \emph{Good}, \emph{Moderate}
and \emph{High}.  These are used only to colour the plot with \emph{Bad} red
and \emph{High} blue.  If missing, regularly spaced boundaries of
\code{c(0.2, 0.4, 0.6, 0.8)} are used with a warning.  If \code{NULL}, the
probability that defines the single \acronym{EQR} is coloured blue.}

\item{mu}{a matrix of posterior samples of the abundance component
\eqn{\mu}{mu} can optionally be given to save recalculation if already
available.  This is assumed to have been calculated from
\code{\link{abundance}} using the same arguments as above.}

\item{rho}{a matrix of posterior samples of the prevalence component
\eqn{\rho}{rho} can optionally be given to save recalculation if already
available.  This is assumed to have been calculated from
\code{\link{prevalence}} using the same arguments as above.}

\item{xmax}{an optional x-axis upper limit giving the largest value of the
total catch to plot the probability of.}

\item{xmaxscale}{if \code{xmax} is not specified, the x-axis upper limit is
set to \code{xmaxscale *} the expected total catch.}

\item{maxpts}{the maximum number of points to plot on the x-axis. If
\code{xmax >= maxpts} a simpler plot with lines rather than bars is drawn
with \code{maxpts} points only.}

\item{title}{an optional character vector giving the title for each plot.
Should have one entry per survey selected by \code{subset}.}

\item{\dots}{further arguments passed to \code{\link{barplot}} if \code{xmax
< maxpts} of \code{\link{plot}} otherwise.}
}
\description{
Plots a probability mass function representing the predicted probability of
catching a given number of fish in total over all runs.  The plot is
coloured to indicate the mean \acronym{EQR} achieved if each number was
observed in terms of the \acronym{WFD} classification this would indicate.
}
\seealso{
\code{\link{fcs2FitModel}} for fitting the model.\cr
\code{\link{pCatch}} for obtaining the probabilities plotted by this
function.\cr

\code{\link{fcs2InteractivePrediction}} can be used to interactively view
how the predicted catch distribution changes with covariates.
}
\keyword{hplot}
