% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppplot.R
\name{ppplot}
\alias{ppplot}
\title{Probability-Probability Plot of FCS2 Model Fit}
\usage{
ppplot(fit, dataFrame, subset = 1:nrow(dataFrame), na.action,
  n.sims = 100, ciprob = 0.95, title = "", addCI = TRUE,
  progressBar = TRUE, mu, rho)
}
\arguments{
\item{fit}{an \code{"fcs2Fit"} object, usually returned by
\code{\link{fcs2FitModel}}.}

\item{dataFrame}{a data frame with surveys as rows and variables as columns.
It should contain all variables required by \code{fit}.  This is usually the
same data frame used to create \code{fit}.}

\item{subset}{an optional vector specifying a subset of surveys to be used
to create the \acronym{P-P} plot.}

\item{na.action}{a function which indicates what should happen when the data
contain missing values (\code{NA}s).  The default is set by the
\code{na.action} setting of \code{\link{options}} and this is usually set to
\code{\link{na.omit}}.  This setting removes surveys that contain missing
data in any required variables.  A vector indicating the rows that were
removed can be extracted from the returned object using
\code{\link{na.action.fcs2Fit}}.  Alternatively, \code{\link{na.pass}} can
be used to ignore missing values (where possible) or \code{\link{na.fail}}
can be given to signal an error if missing values are found.}

\item{n.sims}{the number of datasets to simulate from the fitted model to
generate the confidence interval if \code{addCI} is \code{TRUE}.}

\item{ciprob}{the desired limits for the confidence interval added if
\code{addCI} is \code{TRUE}.  The default is \code{0.95} representing a 95\%
interval.}

\item{title}{an optional title for the plot.}

\item{addCI}{whether to add a confidence interval to the plot (default is
\code{TRUE}).}

\item{progressBar}{whether to show a progress bar when calculating the
confidence interval, since this can take some time.}

\item{mu}{a matrix of abundance samples can optionally be provided, as
generated from \code{\link{abundance}} using the same \code{fit},
\code{dataFrame} and \code{subset}.}

\item{rho}{a matrix of prevalence samples can optionally be provided, as
generated from \code{\link{prevalence}} using the same \code{fit},
\code{dataFrame} and \code{subset}.}
}
\value{
If \code{addCI} is \code{TRUE}, the proportion of observed
\acronym{P-P} plot points within the confidence limits is printed to screen
and invisibly returned.
}
\description{
Produces a \dfn{Probability-Probability} (\acronym{P-P}) plot of the fitted
total catch that can be used to assess the model fit.
}
\details{
If \code{addCI} is \code{TRUE}, a confidence interval is found by Monte
Carlo simulation.  \code{n.sims} datasets of fish catches are simulated from
the fitted model, each using the same observed covariates, and the
\acronym{P-P} plot line is found for each.  The confidence interval then
represents the extent covered by the central \code{ciprob} proportion of
these.

If the observed data is typical of data simulated from the model, we would
expect the observed \acronym{P-P} plot line to fall within the confidence
interval \code{ciprob * 100}\% of the time.
}
\section{Warning}{
 Generating the confidence interval can take some time,
especially if \code{n.sims} is large.
}

\seealso{
\code{\link{plot.fcs2Fit}}
}
\keyword{hplot}
