% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fcs2EQR.R
\name{print.fcs2EQR}
\alias{print.fcs2EQR}
\title{Printing EQR Samples}
\usage{
\method{print}{fcs2EQR}(x)
}
\arguments{
\item{x}{an object of class \code{"fcs2EQR"} containing Monte Carlo
\acronym{EQR} samples, as calculated from \code{\link{fcs2SingleEQR}},
\code{\link{fcs2JointEQR}} or \code{\link{fcs2JointAndSingleEQR}}.}
}
\value{
invisibly returns the \code{"fcs2EQR"} object \code{x}.
}
\description{
Prints a basic summary of a set of \acronym{EQR} samples.
}
\seealso{
\code{\link{summary.fcs2EQR}} for a more detailed summary;\cr
\code{\link{fcs2SingleEQR}}, \code{\link{fcs2JointEQR}} or
\code{\link{fcs2JointAndSingleEQR}} for generating \acronym{EQR} samples.
}
\keyword{print}
