% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fcs2Fit.R
\name{print.fcs2Fit}
\alias{print.fcs2Fit}
\title{Printing an FCS2 Model Fit}
\usage{
\method{print}{fcs2Fit}(x, inla = is.null(x$bugsFit))
}
\arguments{
\item{x}{an object of class \code{"fcs2Fit"}, usually returned by
\code{\link{fcs2FitModel}}.}

\item{inla}{whether to summarise approximate model fits obtained from
\acronym{INLA}, if available.}
}
\value{
invisibly returns the \code{"fcs2Fit"} object \code{x}.
}
\description{
Prints a basic summary of an \acronym{FCS2} model fit.
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
\code{\link{summary.fcs2Fit}} for a more detailed summary.
\code{\link{fcs2FitModel}} for fitting the \acronym{FCS2} model.
}
\keyword{print}
