% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbinom.R
\name{pzinbinom}
\alias{pzinbinom}
\title{Zinbinom}
\usage{
pzinbinom(q, size, prob, zeroprob, nbmean)
}
\arguments{
\item{q}{q}

\item{size}{size}

\item{prob}{prob}

\item{zeroprob}{zero prob}

\item{nbmean}{nbmean}
}
\description{
Distributions.
}
