% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.fcs2EQR.R
\name{quantile.fcs2EQR}
\alias{quantile.fcs2EQR}
\title{Quantiles of an EQR Sample}
\usage{
\method{quantile}{fcs2EQR}(x, probs = seq(0, 1, 0.25), ...)
}
\arguments{
\item{x}{an object of class \code{"fcs2EQR"} containing Monte Carlo
\acronym{EQR} samples, as calculated from \code{\link{fcs2SingleEQR}},
\code{\link{fcs2JointEQR}} or \code{\link{fcs2JointAndSingleEQR}}.}

\item{probs}{numeric vector of probabilities with values in [0, 1].}

\item{\dots}{further arguments passed to \code{\link{quantile}}.}
}
\value{
Returns a matrix or array of quantile estimates for each of the
surveys (or sites/etc if surveys were joined) and species contained in
\code{x}.
}
\description{
Estimates quantiles of an \acronym{EQR} variable from a collection of Monte
Carlo \acronym{EQR} samples.
}
\seealso{
\code{\link{mean.fcs2EQR}}
}
