% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbinom.R
\name{qzinbinom}
\alias{qzinbinom}
\title{Zinbinom}
\usage{
qzinbinom(p, size, prob, zeroprob, nbmean)
}
\arguments{
\item{p}{p}

\item{size}{size}

\item{prob}{prob}

\item{zeroprob}{zero prob}

\item{nbmean}{nbmean}
}
\description{
Distributions.
}
