% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialTerms.R
\name{rw1}
\alias{rw1}
\alias{specialTerms}
\alias{rw2}
\alias{spatial}
\title{Special Non-Linear Terms for FCS2 Abundance and Prevalence Regressions}
\usage{
rw1(val, noLevels = 10, boundaries, scale.parameters,
  precision.parameters, initial.values, precision.initial.value)
}
\arguments{
\item{val}{a covariate to use for the non-linear term.  \code{rw1} and
\code{rw2} assume a continuous variable but \code{spatial} requires an
integer variable that indicates which spatial region every survey is
within.}

\item{noLevels}{the number of discrete levels to use to represent the
continuous non-linear relationship. This is ignored if \code{boundaries} is
supplied.}

\item{boundaries}{an optional vector specifying the location of the discrete
points that are used to represent the non-linear covariate term. If
missing, \code{noLevels} boundaries are created spaced regularly between
the minimum and maximum values of \code{val}.}

\item{scale.parameters}{an optional vector of length 2 specifying the prior
distribution of the scale hyperparameter (\eqn{\sigma} or \eqn{\nu}) that
controls the variability between levels in the random walk or adjacent
spatial regions.}

\item{precision.parameters}{an optional vector of length 2 specifying the
prior distribution of the precision hyperparameter (\eqn{\tau = 1 /
\sigma^2} or \eqn{\phi = 1 / \nu^2}) that controls the variability between
levels in the random walk or adjacent spatial regions. This is ignored if
\code{scale.parameters} is specified.}

\item{initial.values}{an optional vector giving an estimate for each of the
variables that make up the term, one for each random walk level or one for
each spatial region. These are used as starting values for the
\acronym{MCMC} chains when fitting the full model using \acronym{BUGS}.}

\item{precision.initial.value}{an optional starting value for the precision
hyperparameter to be used when fitting the full model using \acronym{BUGS}.}

\item{adjacency}{a list with three components containing adjacency
information relating to the spatial region. The first component \code{num}
is a vector containing the number of regions adjacent to each region. The
second component \code{adj} is a vector listing the indices of each of
these adjacent regions. The third component \code{sumNumNeigh} is the sum
of the number of neighbours, which should equal \code{sum(num)} and
\code{length(adj)}.\cr This list will usually be generated externally from
\R by the adjacency tool within WinBUGS or OpenBUGS. The list should be
given a name and then loaded into \R.}
}
\description{
Non-linear random walk and spatial terms that can be used in addition to
linear terms in the \acronym{FCS2} abundance and prevalence regression
equations. These functions can be used within the formulae when specifying the
abundance or prevalence regression equations via the arguments
\code{muFormula} and \code{rhoFormula} of \code{\link{fcs2FitModel}}.
}
\details{
These functions can be used within the \R formulae when specifying
  the abundance or prevalence regression equations via the arguments
  \code{muFormula} and \code{rhoFormula} of \code{\link{fcs2FitModel}}.
}
\seealso{
\code{\link{fcs2FitModel}}
}
