% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbinom.R
\name{rzinbinom}
\alias{rzinbinom}
\title{Zinbinom}
\usage{
rzinbinom(n, size, prob, zeroprob, nbmean)
}
\arguments{
\item{n}{n}

\item{size}{size}

\item{prob}{prob}

\item{zeroprob}{zero prob}

\item{nbmean}{nbmean}
}
\description{
Distributions.
}
