% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbinom.R
\name{rzinbinom_constrained}
\alias{rzinbinom_constrained}
\title{Zinbinom}
\usage{
rzinbinom_constrained(n, size, prob, zeroprob, nbmean, min = 0,
  max = Inf)
}
\arguments{
\item{n}{n}

\item{size}{size}

\item{prob}{prob}

\item{zeroprob}{zero prob}

\item{nbmean}{nbmean}

\item{min}{min}

\item{max}{max}
}
\description{
Distributions.
}
