% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SalmonFryFit.R
\docType{data}
\name{SalmonFryFit}
\alias{SalmonFryFit}
\title{SalmonFryFit}
\format{A data frame.}
\usage{
data(SalmonFryFit)
}
\description{
A dataset containing model data.
}
\examples{
\dontrun{data(SalmonFryFit)}
}
\keyword{datasets}
