% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SalmonParrFit.R
\docType{data}
\name{SalmonParrFit}
\alias{SalmonParrFit}
\title{SalmonParrFit}
\format{A data frame.}
\usage{
data(SalmonParrFit)
}
\description{
A dataset containing model data.
}
\examples{
\dontrun{data(SalmonParrFit)}
}
\keyword{datasets}
