% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fcs2EQR.R
\name{summary.fcs2EQR}
\alias{summary.fcs2EQR}
\title{Detailed EQR summary}
\usage{
\method{summary}{fcs2EQR}(object)
}
\arguments{
\item{object}{summary object}
}
\value{
detailed EQR summary object
}
\description{
Provides a more detailed EQR summary
}
\examples{
\dontrun{
summary.fcs2EQR(object = data)
}
}
