% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fcs2Fit.R
\name{summary.fcs2Fit}
\alias{summary.fcs2Fit}
\title{Detailed summary}
\usage{
\method{summary}{fcs2Fit}(object, prior = TRUE,
  inla = is.null(object$bugsFit), posterior = !is.null(object$bugsFit),
  allVars = FALSE, ...)
}
\arguments{
\item{object}{summary object}

\item{prior}{from summary object}

\item{inla}{from summary object}

\item{posterior}{from summary object}

\item{allVars}{boolean}

\item{\dots}{Not currently used.}
}
\value{
detailed summary object
}
\description{
Provides a more detailed summary
}
\examples{
\dontrun{
summary.fcs2Fit(object = data)
}
}
\seealso{
summary.fcs2Fit summary.fcs2Fit print.summary.fcs2Fit
}
