% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinBUGSSamples.R
\name{thinBUGSSamples}
\alias{thinBUGSSamples}
\title{Thin BUGS Samples}
\usage{
thinBUGSSamples(fit, n.thin = 1, first = 1,
  last = fit$bugsFit$n.keep, which, chains = 1:fit$bugsFit$n.chains)
}
\arguments{
\item{fit}{fit}

\item{n.thin}{number to thin}

\item{first}{first}

\item{last}{last}

\item{which}{which}

\item{chains}{chains}
}
\value{
thinner samples
}
\description{
Used when necessary with a warning to thin samples so that the number per
chain is consistent across fits.
}
\examples{
\dontrun{
thinBUGSSamples(fit,
                n.thin = 1,
                first = 1,
                last = fit$bugsFit$n.keep,
                which,
                chains = 1:fit$bugsFit$n.chains)
}
}
