% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TroutParrFit.R
\docType{data}
\name{TroutParrFit}
\alias{TroutParrFit}
\title{TroutParrFit}
\format{A data frame.}
\usage{
data(TroutParrFit)
}
\description{
A dataset containing model data.
}
\examples{
\dontrun{data(TroutParrFit)}
}
\keyword{datasets}
