% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.names.fcs2Fit.R
\encoding{UTF-8}
\name{variable.names}
\alias{variable.names}
\alias{variable.names.fcs2Fit}
\title{Variable Names in FCS2 Model}
\usage{
\method{variable.names}{fcs2Fit}(
  object,
  q = object$multiRun,
  r = TRUE,
  abundance = TRUE,
  prevalence = TRUE,
  linear = TRUE,
  rw = "group",
  spatial = "group",
  hyperparams = "scale",
  ...
)
}
\arguments{
\item{object}{an object of class \code{"fcs2Fit"}, usually returned by \code{\link{fcs2FitModel}}.}

\item{q}{whether to include the catch probability variable \eqn{q}.}

\item{r}{whether to include the shape parameter \eqn{r} (given by \code{size}
in \code{\link{dzinbinom}} and \code{\link{dzinmultinom}}).}

\item{abundance}{whether to include any variables related to the abundance regression.}

\item{prevalence}{whether to include any variables related to the prevalence regression.}

\item{linear}{whether to include linear regression terms.}

\item{rw}{whether and how to include random walk terms. If \code{TRUE} or \code{"group"},
a single variable name is returned for each random walk term. If \code{"singluar"},
each individual variable is included.}

\item{spatial}{whether and how to include spatial terms. If \code{TRUE} or \code{"group"},
a single variable name is returned for each spatial term. If \code{"singluar"},
each individual variable is included.}

\item{hyperparams}{whether and how to include hyperparameters. If \code{"scale"},
scale hyperparameters (\eqn{\sigma} or \eqn{\nu}) are included. If \code{"precision"},
hyperparameters are specified on the precision scale (\eqn{\tau} or \eqn{\phi}).}

\item{\dots}{Not currently used.}
}
\value{
A character vector of selected variables in the \acronym{FCS2} model specified by \code{object}.
}
\description{
Returns a vector of selected variables in an \acronym{FCS2} model.
}
\seealso{
\code{\link{fcs2FitModel}} for fitting the \acronym{FCS2} model.
}
