% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbinom.R
\encoding{UTF-8}
\name{zinbinom}
\alias{zinbinom}
\alias{dzinbinom}
\alias{dzimbinom}
\alias{pzinbinom}
\alias{qzinbinom}
\alias{rzinbinom}
\alias{rzinbinom_constrained}
\title{Zero-inflated Negative Binomial Distributions.}
\usage{
dzinbinom(x, size, prob, zeroprob, nbmean)

pzinbinom(q, size, prob, zeroprob, nbmean)

qzinbinom(p, size, prob, zeroprob, nbmean)

rzinbinom(n, size, prob, zeroprob, nbmean)

rzinbinom_constrained(n, size, prob, zeroprob, nbmean, min = 0,
  max = Inf)
}
\arguments{
\item{x}{x}

\item{size}{size}

\item{prob}{prob}

\item{zeroprob}{zero prob}

\item{nbmean}{nbmean}

\item{q}{q}

\item{p}{p}

\item{n}{n}

\item{min}{min}

\item{max}{max}
}
\value{
zero-inflated Negative Binomial Distributions.
}
\description{
Zero-inflated Negative Binomial Distributions.
}
\examples{
\dontrun{
 data$Catch <- rzinbinom(100,
size = 1.1, zeroprob = 0.3,
nbmean = 0.3 * Data$SurveyArea
)
}
}
