% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinmultinom.R
\encoding{UTF-8}
\name{zinmultinom}
\alias{zinmultinom}
\alias{dzinmultinom}
\alias{rzinmultinom}
\title{Zero-inflated Negative Multinomial Distributions.}
\usage{
dzinmultinom(x, size, prob, zeroprob, nmmean)

rzinmultinom(n, size, prob, zeroprob, nmmean)

rzinmultinom(n, size, prob, zeroprob, nmmean)
}
\arguments{
\item{x}{x}

\item{size}{size}

\item{prob}{prob}

\item{zeroprob}{zero prob}

\item{nmmean}{nmmean}

\item{n}{n}
}
\value{
zero-inflated Negative Multinomial Distributions.
}
\description{
Zero-inflated Negative Multinomial Distributions.
}
