% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TroutFryFit.R
\docType{data}
\name{TroutFryFit}
\alias{TroutFryFit}
\title{TroutFryFit}
\format{A data frame.}
\usage{
data(TroutFryFit)
}
\description{
A dataset containing model data.
}
\examples{
\dontrun{data(TroutFryFit)}
}
\keyword{datasets}
